refactor: remove one definition of buildProteusService
