+++ packages/core/src/messagingProtocols/proteus/ProteusService/ProteusService.mocks.ts
@@ -17,34 +17,23 @@
  *
  */
 
-import {ClientClassification, ClientType} from '@wireapp/api-client/lib/client';
-import {CoreCrypto} from '@wireapp/core-crypto/platforms/web/corecrypto';
+import {ClientType} from '@wireapp/api-client/lib/client';
 
 import {APIClient} from '@wireapp/api-client';
+import {CoreCrypto} from '@wireapp/core-crypto';
+import {MemoryEngine} from '@wireapp/store-engine';
 
 import {ProteusService} from './ProteusService';
 
-import {MemoryEngine} from '../../../../../store-engine/lib';
 import {CryptographyService} from '../../../cryptography';
 import {getUUID} from '../../../test/PayloadHelper';
 
-export const buildProteusService = (
+export const buildProteusService = async (
   federated = false,
-): [ProteusService, {apiClient: APIClient; coreCrypto: CoreCrypto; cryptographyService: CryptographyService}] => {
+): Promise<
+  [ProteusService, {apiClient: APIClient; coreCrypto: CoreCrypto; cryptographyService: CryptographyService}]
+> => {
   const apiClient = new APIClient({urls: APIClient.BACKEND.STAGING});
-  jest.spyOn(apiClient.api.user, 'postListClients').mockImplementation(() =>
-    Promise.resolve({
-      qualified_user_map: {
-        'test-domain': {
-          'test-id-1': [{class: ClientClassification.DESKTOP, id: 'test-client-id-1-user-1'}],
-          'test-id-2': [
-            {class: ClientClassification.DESKTOP, id: 'test-client-id-1-user-2'},
-            {class: ClientClassification.PHONE, id: 'test-client-id-2-user-2'},
-          ],
-        },
-      },
-    }),
-  );
 
   apiClient.context = {
     clientType: ClientType.NONE,
@@ -52,21 +41,15 @@ export const buildProteusService = (
     clientId: getUUID(),
   };
 
+  const coreCrypto = await CoreCrypto.deferredInit('store-name', 'key');
+
   const cryptographyService = new CryptographyService(apiClient, new MemoryEngine(), {
-    useQualifiedIds: false,
+    useQualifiedIds: federated,
     nbPrekeys: 1,
   });
 
-  const coreCrypto = {
-    getRemoteFingerprint: jest.fn(),
-    getLocalFingerprint: jest.fn(),
-    proteusSessionExists: jest.fn(),
-    proteusSessionFromPrekey: jest.fn(),
-    proteusFingerprintRemote: jest.fn(),
-  } as unknown as CoreCrypto;
-
-  return [
-    new ProteusService(apiClient, cryptographyService, coreCrypto, {useQualifiedIds: federated}),
-    {apiClient, coreCrypto, cryptographyService},
-  ];
+  const proteusService = new ProteusService(apiClient, cryptographyService, coreCrypto, {
+    useQualifiedIds: federated,
+  });
+  return [proteusService, {apiClient, coreCrypto, cryptographyService}];
 };
