+++ packages/core/src/messagingProtocols/proteus/ProteusService/Utility/getSessionsAndClientsFromRecipients/getSessionsAndClientsFromRecipients.test.ts
@@ -21,7 +21,7 @@ import {ClientClassification} from '@wireapp/api-client/lib/client';
 import type {UserClients} from '@wireapp/api-client/lib/conversation';
 import type {UserPreKeyBundleMap} from '@wireapp/api-client/lib/user';
 
-import {buildProteusService} from '../../../../../test/ProteusHelper';
+import {buildProteusService} from '../../ProteusService.mocks';
 import {preKeyBundleToUserClients} from '../preKeyBundleToUserClients';
 
 import {getSessionsAndClientsFromRecipients} from './';
